NOX-245 - Support Localized Text Fields for Entities with AutoNumber Key (#1120)
* added localization support for entities with autoNumber id;

* added many-to-many relationship tests to tenant; skipping failing many-to-many relationship tests on workplace;

* resolved PR comments;

---------

+++ src/Nox.ClientApi.Tests/Tests/Controllers/CountriesControllerTests.cs
@@ -774,7 +774,7 @@ namespace ClientApi.Tests.Tests.Controllers
                 new WorkplaceCreateDto() { Name = _fixture.Create<string>() });
             var countryCreateDto = new CountryCreateDto { 
                 Name = _fixture.Create<string>(),
-                WorkplacesId = new List<UInt32> {
+                WorkplacesId = new List<long> {
                     workplaceResponse1!.Id,
                     workplaceResponse2!.Id,
                     workplaceResponse3!.Id,
@@ -803,7 +803,7 @@ namespace ClientApi.Tests.Tests.Controllers
             var countryCreateDto = new CountryCreateDto
             {
                 Name = _fixture.Create<string>(),
-                WorkplacesId = new List<UInt32> { _fixture.Create<UInt32>() }
+                WorkplacesId = new List<long> { _fixture.Create<long>() }
             };
 
             // Act
